Simplified prepare-to-send durable function.
